resonator masking, new iq plotting/tunning, resonator tone writing
I added in new interative plotting of iq sweeps that also can be
used to retune the tone to the minimum maginitude or the maximum
in didq/df. I added in a function mask which resonators are
currently readout based on a frequency separation. I added in a
function to write the tones for non masked resonators. I moved
merge into set_frequencies where it belongs not in vna_sweep and
iq_sweeps. updated iq_sweep_freqs to be actual frequencies not
just lo frequencies