more iq plotting, fit iq sweeps, power tuning for resonators
added in additional feature to the interactive iq plotting. Now
it can display data for all resoantors in a third subplot. This
is used to display the results of fits to the iq sweeps. Added in
some basic fitting of the iq sweeps mostly to get information out
about the non-linearity paramter a so that we can tune the tone
powers correctly. added in an interactive plot for tuning the tone
powers. Also fixed a bug for data_io of filename set reading